test: mesclar lead-logic.test.ts em lib-lead-logic.test.ts
tests/lead-logic.test.ts era um subconjunto de lib-lead-logic.test.ts,
ambos testando cleanString/normalizeNullable de lib/lead-logic.ts.
Move os casos únicos (cleanString completo + truncamento/XSS de
normalizeNullable) para o destino e remove o arquivo duplicado.
